pyproject.toml
[project]
name = "warning-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Pugachev-Mihail",email = "xd383727@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.19.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "motor (>=3.7.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "warning_bot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
